Force cast cell.item instead of optional failure
cell.item should NEVER be nil and it being so is a case of very bad
logical error occuring somewhere. One of the danger of optionals is
always taking the easy route of letting it pass as a ? and failing
silently. You'd want this to fail immediately so you caught it during
development.

Perhaps a better option would be

    assert(cell.item != nil)
    let item = cell.item ?? Style(name: "7s") // or some other sane default